+++ app/widgets/Chat/Chat.php
@@ -1088,6 +1088,13 @@ class Chat extends \Movim\Widget\Base
             $message->body = $message->subject .': '. $message->body;
         }
 
+        // XEP-0393
+        // $message->body = (preg_replace ('/```((.|\n)*?)(```|\z)/', "<pre>$1</pre>", $message->body));
+        $message->body = (preg_replace ('/(`(?!`).*?`)/', "<code>$1</code>", $message->body));
+        $message->body = (preg_replace ('/(\*.*?\*)/', "<b>$1</b>", $message->body));
+        $message->body = (preg_replace ('/(_.*?_)/', "<em>$1</em>", $message->body));
+        $message->body = (preg_replace ('/(~.*?~)/', "<s>$1</s>", $message->body));
+
         // Sticker message
         if (isset($message->sticker)) {
             $p = new Picture;